Ascii out
